refactored individual test cases by adding testutils
